# apps/api/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Anplexa API",
    "version": "1.0.0",
...
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate with email and password to receive JWT tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Exchange a valid refresh token for a new access token pair.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New tokens returned" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "funnelAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Funnel API key"
      }
    },
...
  }
}